refactor dir names
