+++ src/app/layout/index.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react';
 import styles from './styles.module.scss';
-import { Header } from '@widgets/Header';
+import { Header } from '@widgets/header';
 
 interface ILayout {
     children: ReactNode;
