--- src/features/LikeProduct/index.ts
@@ -1,1 +1,0 @@
-export { LikeProduct } from './ui/'
