+++ src/features/like-product/index.ts
@@ -1,0 +1,1 @@
+export { LikeProduct } from './ui'
