+++ src/pages/Home/ui/about-mobile/styles.module.scss
