+++ src/pages/Home/ui/hero-section/styles.module.scss
