+++ src/pages/Home/ui/top-restaurants/index.tsx
