+++ src/shared/ui/custom-icon/index.tsx
