+++ src/shared/ui/custom-icon/styles.module.scss
