+++ src/shared/ui/custom-icon/types.ts
