+++ src/shared/ui/separator-line/styles.module.scss
