+++ src/widgets/Header/ui/app-logo/index.tsx
