+++ src/widgets/Header/ui/app-logo/styles.module.scss
